refactor(website): Move Home-related files to a separate folder
